Changed the window initialization code from utils to core
