+++ src/core/mod.rs
@@ -2,7 +2,9 @@ pub mod buffers;
 pub mod shader;
 pub mod traits;
 pub mod transform;
+pub mod window;
 
-pub use shader::Shader;
-pub use traits::{Drawable, Transformable};
-pub use transform::Transform;
+pub use shader::*;
+pub use traits::*;
+pub use transform::*;
+pub use window::*;
