+++ src/lib.rs
@@ -2,3 +2,5 @@ pub mod components;
 pub mod core;
 pub mod io;
 pub mod utils;
+
+pub use crate::core::*;
